# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Valheim Server UI API
  version: 1.0.0
  description: |
    HTTP API of valheim-server-ui. All endpoints are under /api/v1.
    Roles: viewer < operator < admin. Each operation declares the minimum role in `x-role`.
    Mutating requests must send header `X-Requested-With: valheim-ui` (CSRF guard).
    Long-running operations return 202 with a Job; follow progress via /events (SSE).
servers:
  - url: /api/v1
...
paths:
...
  /instances/{instanceId}/worlds:
    parameters: [{ $ref: '#/components/parameters/instanceId' }]
    get:
      tags: [worlds]
      x-role: viewer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                required: [worlds]
                properties:
                  worlds: { type: array, items: { $ref: '#/components/schemas/World' } }
    post:
      tags: [worlds]
      summary: Upload a world (.db + .fwl files, or one zip containing them)
      x-role: operator
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files: { type: array, items: { type: string, format: binary } }
                overwrite: { type: boolean, default: false }
      responses:
        '202': { $ref: '#/components/responses/JobResponse' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /instances/{instanceId}/worlds/{worldName}:
    parameters:
      - { $ref: '#/components/parameters/instanceId' }
      - { $ref: '#/components/parameters/worldName' }
    delete:
      tags: [worlds]
      summary: Delete an inactive world
      x-role: operator
      responses:
        '204': { description: Deleted }
        '409': { $ref: '#/components/responses/Error' }
  /instances/{instanceId}/worlds/{worldName}/regenerate:
    post:
      tags: [worlds]
      summary: |
        Regenerate the active world: take a backup, delete its save files and let
        Valheim create a fresh world (new random seed) with the same name on the
        next start. Only the active world can be regenerated; delete inactive ones.
      x-role: operator
      parameters:
        - { $ref: '#/components/parameters/instanceId' }
        - { $ref: '#/components/parameters/worldName' }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StopIfRunning' }
      responses:
        '202': { $ref: '#/components/responses/JobResponse' }
        '409': { $ref: '#/components/responses/Error' }
  /instances/{instanceId}/worlds/{worldName}/download:
    get:
      tags: [worlds]
      x-role: operator
      parameters:
        - { $ref: '#/components/parameters/instanceId' }
        - { $ref: '#/components/parameters/worldName' }
      responses:
        '200':
          content:
            application/zip: { schema: { type: string, format: binary } }

  # ------------------------------------------------------------------ backups
...
components:
  parameters:
    instanceId: { in: path, name: instanceId, required: true, schema: { type: string, pattern: '^[a-z0-9][a-z0-9-]{0,31}$' } }
...
    worldName: { in: path, name: worldName, required: true, schema: { type: string } }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    JobResponse:
      description: Job
      content:
        application/json:
          schema:
            type: object
            required: [job]
            properties:
              job: { $ref: '#/components/schemas/Job' }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                One of: unauthorized, forbidden, not_found, validation_failed, conflict,
                instance_running, instance_not_installed, instance_busy, port_in_use,
                last_admin, invalid_credentials, account_locked, account_disabled,
                setup_done, oidc_disabled, oidc_error, steamcmd_missing, bepinex_missing,
                package_not_found, upstream_error, internal
            message: { type: string }
            details:
              type: object
              additionalProperties: true
              properties:
                fields:
                  type: array
                  items:
                    type: object
                    required: [field, message]
                    properties:
                      field: { type: string }
                      message: { type: string }
...
    StopIfRunning:
      type: object
      properties:
        stop_if_running: { type: boolean, default: false, description: Stop the instance for the operation and start it again afterwards }
...
    World:
      type: object
      required: [name, active, size_bytes, modified_at]
      properties:
        name: { type: string }
        active: { type: boolean }
        size_bytes: { type: integer, format: int64 }
        modified_at: { type: string, format: date-time }
        has_db: { type: boolean }
        has_fwl: { type: boolean }
...
    JobType:
      type: string
      enum: [install, update, backup, restore, world_import, world_regenerate, mod_install, mod_update, mod_uninstall, bepinex_install, scheduled_restart, thunderstore_refresh, self_upgrade]
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
    Job:
      type: object
      required: [id, type, status, created_at]
      properties:
        id: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/JobType' }
        instance_id: { type: string }
        status: { $ref: '#/components/schemas/JobStatus' }
        title: { type: string, description: Human readable, e.g. "Install Jotunn 2.29.2" }
        requested_by: { type: string }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        error: { type: string }
        summary:
          type: object
          additionalProperties: true
...